Change from make machinery for compilation to CMake which also enables coupling to the smash libraries.
For this, the file structure needed to be slightly modified, such that
there is now a separate directory within the src directory that
contains only the headers (called src/include).
Also, the Makefile was deleted and basically replaced by a CMake file.
Not that for proper linking to the smash libraries it is necessary to
follow the instructions on
https://github.com/smash-transport/smash/tree/master/examples/using_SMASH_as_library.