Provide Oscar2013 format as an output
In addition to the ROOT output format, the ASCII Oscar2013 format is
provided as an additional output. It relies on the smash Oscar2013
format (not extended).
By default, both the ROOT and the Oscar output are now generated upon
execution. For simplicity, the entire implementation is done in
the corresponding header.

Further changes/improvements:
* Include statements are cleaned up and sorted alphabetically
* Include guards are introduced to prevent re-declarations of class
members
* ROOT final state tree is removed as it is identical to the initial
state (Particle decays are not performed)