Increase muQ-cutoff to 0.13
This cutoff is necessary to not have mu > mass, where mu is the full
chemical potential. In the case of mu > mass, the particle cannot be
sampled. As such this cutoff serves as a safeguard. From experience, the
chemical potential is however usually below this value. Only rarely it
can have values around +- 0.15.