Update sampler to be compatible with latest SMASH version
The output functions in smash were changed in commit
be4b7dd7a3a3765e3b64ae29bf063fbae6206ea3. The event_info struct was
introduced to hold multiple paramaters and passes the necessary information to the output
functions. The information stored therein is not necessary in the
context of the sampler, so the struct is filled with dummy parameters.